refactor: quiz question as a component
